Сcorrected comments
